project1-fraud-detection/requirements.txt
faker==24.0.0
pandas==2.2.0
pyarrow==15.0.0
pyspark==3.5.0
apache-airflow==2.8.0
apache-airflow-providers-google==10.13.0
apache-airflow-providers-slack==8.5.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
google-cloud-bigquery==3.17.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.20.0
tenacity==8.2.3
pytest==7.4.4
